refactor: decouple jq package from k8s.io/apimachinery
Move Unstructured converters from jq to k8s package, registered via
init(). The jq package no longer imports k8s.io/apimachinery, making
it usable without pulling in the Kubernetes dependency tree.
